Issues-71: Create separate module for first example of DiffUtils.
