+++ app/build.gradle
@@ -62,4 +62,5 @@ dependencies {
     implementation project(path: ':simpletransition')
     implementation project(path: ':likebutton')
     implementation project(path: ':recyclercollapsing')
+    implementation project(path: ':timediffutils')
 }
