--- app/src/main/java/com/example/kirchhoff/example/ui/recycler/diffutil/ex1/Time.kt
@@ -1,6 +1,0 @@
-package com.example.kirchhoff.example.ui.recycler.diffutil.ex1
-
-/**
- * @author Kirchhoff-
- */
-data class Time(val id: String, val hours: Int, val minute: Int, val seconds: Int)
