+++ timediffutils/src/main/java/com/example/timediffutils/Time.kt
@@ -1,0 +1,3 @@
+package com.example.timediffutils
+
+data class Time(val id: String, val hours: Int, val minute: Int, val seconds: Int)
