+++ timediffutils/src/main/java/com/example/timediffutils/extensions/ViewExtensions.kt
@@ -1,10 +1,7 @@
-package com.example.kirchhoff.example.ui.extensions
+package com.example.timediffutils.extensions
 
 import android.widget.TextView
 
-/**
- * @author Kirchhoff-
- */
 fun TextView.setFormatDigit(digit: Int) {
     text = String.format("%02d", digit)
 }
