+++ timediffutils/src/main/res/layout/r_time_diff.xml
