+++ timediffutils/src/main/res/values/arrays.xml
